backend/requirements.txt
fastapi
uvicorn
python-jose
passlib[argon2]
groq
python-dotenv
sqlalchemy
python-multipart
pydantic[email]
